WIP Commit
added comments to some analytics places and updated the hint text for the phone based login.
tics the release to 20.6
The .pbxproj is also modified and could contain artifacts of an issue with main.jsbundle missing image assets (path changes to ios/assets/assets). This should be verified fixed before a merge.